tests/wpt/tests/html/canvas/element/filters/2d.filter.canvasFilterObject.componentTransfer.linear.tentative.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.filter.canvasFilterObject.componentTransfer.linear.tentative</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.filter.canvasFilterObject.componentTransfer.linear.tentative</h1>
<p class="desc">Test pixels on CanvasFilter() componentTransfer with linear type</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Test pixels on CanvasFilter() componentTransfer with linear type");
_addTest(function(canvas, ctx) {

  // From https://www.w3.org/TR/SVG11/filters.html#feComponentTransferElement
  function getColor(inputColor, slopes, intercepts) {
      return [
          Math.max(0, Math.min(1, inputColor[0]/255 * slopes[0] + intercepts[0])) * 255,
          Math.max(0, Math.min(1, inputColor[1]/255 * slopes[1] + intercepts[1])) * 255,
          Math.max(0, Math.min(1, inputColor[2]/255 * slopes[2] + intercepts[2])) * 255,
      ];
  }

  const slopes = [0.5, 1.2, -0.2];
  const intercepts = [0.25, 0, 0.5];
  ctx.filter = new CanvasFilter({name: 'componentTransfer',
      funcR: {type: 'linear', slope: slopes[0], intercept: intercepts[0]},
      funcG: {type: 'linear', slope: slopes[1], intercept: intercepts[1]},
      funcB: {type: 'linear', slope: slopes[2], intercept: intercepts[2]},
  });

  const inputColors = [
      [255, 255, 255],
      [0, 0, 0],
      [127, 0, 34],
      [252, 186, 3],
      [50, 68, 87],
  ];

  for (const color of inputColors) {
      let outputColor = getColor(color, slopes, intercepts);
      ctx.fillStyle = `rgb(${color[0]}, ${color[1]}, ${color[2]})`;
      ctx.fillRect(0, 0, 10, 10);
      _assertPixelApprox(canvas, 5, 5, outputColor[0],outputColor[1],outputColor[2],255, 2);
  }

});
</script>
